Split data-dumping serialisers into their respective apps … so that new ones can easily be auto-discovered and for use in another upcoming management command. Existing serialisers' output is unchanged; only the location they load from.
